# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisInstrumentListResponse": {
        "properties": {
          "instruments": {
            "items": {
              "$ref": "#/components/schemas/AnalysisInstrumentResponse"
            },
            "title": "Instruments",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "offset",
          "limit",
          "instruments"
        ],
        "title": "AnalysisInstrumentListResponse",
        "type": "object"
      },
      "AnalysisInstrumentResponse": {
        "properties": {
          "base_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Asset"
          },
          "company_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "first_session": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Session"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "instrument_type": {
            "title": "Instrument Type",
            "type": "string"
          },
          "last_session": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Session"
          },
          "price_basis": {
            "title": "Price Basis",
            "type": "string"
          },
          "price_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Source"
          },
          "quote_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote Asset"
          },
          "stored_sessions": {
            "title": "Stored Sessions",
            "type": "integer"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "venue_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue Code"
          },
          "venue_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue Name"
          }
        },
        "required": [
          "id",
          "symbol",
          "instrument_type",
          "currency",
          "price_basis",
          "stored_sessions"
        ],
        "title": "AnalysisInstrumentResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InstrumentPriceHistoryResponse": {
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "fetched_at": {
            "title": "Fetched At",
            "type": "string"
          },
          "first_date": {
            "title": "First Date",
            "type": "string"
          },
          "fundamentals_fields": {
            "items": {
              "type": "string"
            },
            "title": "Fundamentals Fields",
            "type": "array"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "integer"
          },
          "last_date": {
            "title": "Last Date",
            "type": "string"
          },
          "price_basis": {
            "title": "Price Basis",
            "type": "string"
          },
          "prices": {
            "items": {
              "$ref": "#/components/schemas/InstrumentPricePointResponse"
            },
            "title": "Prices",
            "type": "array"
          },
          "provider_ratio_effective_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Ratio Effective Date"
          },
          "provider_ratio_period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Ratio Period"
          },
          "provider_reported_pb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Reported Pb"
          },
          "provider_reported_pe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Reported Pe"
          },
          "relative_strength_benchmark": {
            "enum": [
              "VN30",
              "SPX"
            ],
            "title": "Relative Strength Benchmark",
            "type": "string"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "shares_cagr_5y_observed_years": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares Cagr 5Y Observed Years"
          },
          "shares_cagr_5y_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares Cagr 5Y Pct"
          },
          "shares_cagr_5y_start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares Cagr 5Y Start Date"
          },
          "shares_cagr_full_5y": {
            "default": false,
            "title": "Shares Cagr Full 5Y",
            "type": "boolean"
          },
          "shares_growth_cagr_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares Growth Cagr Pct"
          },
          "shares_growth_full_10y": {
            "default": false,
            "title": "Shares Growth Full 10Y",
            "type": "boolean"
          },
          "shares_growth_observed_years": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares Growth Observed Years"
          },
          "shares_growth_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares Growth Pct"
          },
          "shares_growth_start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares Growth Start Date"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "trailing_pe_fetched_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trailing Pe Fetched At"
          },
          "trailing_pe_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trailing Pe Method"
          },
          "trailing_pe_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trailing Pe Source"
          },
          "venue_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue Code"
          }
        },
        "required": [
          "instrument_id",
          "symbol",
          "currency",
          "source",
          "price_basis",
          "fetched_at",
          "first_date",
          "last_date",
          "row_count",
          "relative_strength_benchmark",
          "prices"
        ],
        "title": "InstrumentPriceHistoryResponse",
        "type": "object"
      },
      "InstrumentPricePointResponse": {
        "properties": {
          "close": {
            "title": "Close",
            "type": "number"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "eps_ttm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eps Ttm"
          },
          "high": {
            "title": "High",
            "type": "number"
          },
          "low": {
            "title": "Low",
            "type": "number"
          },
          "open": {
            "title": "Open",
            "type": "number"
          },
          "relative_strength": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative Strength"
          },
          "shares_outstanding": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares Outstanding"
          },
          "trailing_pb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trailing Pb"
          },
          "trailing_pe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trailing Pe"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          }
        },
        "required": [
          "date",
          "open",
          "high",
          "low",
          "close"
        ],
        "title": "InstrumentPricePointResponse",
        "type": "object"
      },
...
      "ReferenceRateListResponse": {
        "properties": {
          "facets": {
            "$ref": "#/components/schemas/ReferenceRateFacetsResponse"
          },
          "instruments": {
            "items": {
              "$ref": "#/components/schemas/ReferenceRateInstrumentResponse"
            },
            "title": "Instruments",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "summary": {
            "$ref": "#/components/schemas/ReferenceRateSummaryResponse"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "offset",
          "limit",
          "instruments",
          "facets",
          "summary"
        ],
        "title": "ReferenceRateListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backtesting and factor analysis API",
    "title": "Trading Statistics Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/instruments": {
      "get": {
        "operationId": "listAnalysisInstruments",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "equity",
                    "crypto_spot",
                    "reference_rate"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope"
            }
          },
          {
            "in": "query",
            "name": "universe",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Universe"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "has_price_history",
            "required": false,
            "schema": {
              "default": true,
              "title": "Has Price History",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisInstrumentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Analysis Instruments",
        "tags": [
          "instruments"
        ]
      }
    },
    "/instruments/{instrument_id}/history": {
      "get": {
        "operationId": "getInstrumentPriceHistory",
        "parameters": [
          {
            "in": "path",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "title": "Instrument Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentPriceHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Instrument Price History",
        "tags": [
          "instruments"
        ]
      }
    },
    "/reference-rates": {
      "get": {
        "operationId": "listReferenceRates",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "base_asset",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Base Asset"
            }
          },
          {
            "in": "query",
            "name": "quote_asset",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quote Asset"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "active",
              "enum": [
                "active",
                "inactive",
                "all"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceRateListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reference Rates",
        "tags": [
          "reference-rates"
        ]
      }
    },
...
  }
}